fix details by server
