+++ app.js
@@ -1,7 +1,13 @@
 const express = require("express");
 const cors = require("cors");
 const logger = require("morgan");
-const contactsRouter = require("./contacts");
+const {
+  getAllContacts,
+  getContactById,
+  addNewContact,
+  updateContact,
+  deleteContact,
+} = require("./models/contacts");
 
 const app = express();
 
@@ -10,10 +16,16 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cors());
 app.use(logger("dev"));
 
-app.use("/api/contacts", contactsRouter);
+app.get("/contacts", getAllContacts);
+app.get("/contacts/:id", getContactById);
+app.post("/contacts", addNewContact);
+app.put("/contacts/:id", updateContact);
+app.delete("/contacts/:id", deleteContact);
 
-app.listen(3000, () => {
-  console.log("Server is running on port 3000");
+const port = 3004;
+
+app.listen(port, () => {
+  console.log(`Server is running on port ${port}`);
 });
 
 module.exports = app;
