+++ models/contactsOld.js
@@ -1,0 +1,19 @@
+// const fs = require('fs/promises')
+
+const listContacts = async () => {}
+
+const getContactById = async (contactId) => {}
+
+const removeContact = async (contactId) => {}
+
+const addContact = async (body) => {}
+
+const updateContact = async (contactId, body) => {}
+
+module.exports = {
+  listContacts,
+  getContactById,
+  removeContact,
+  addContact,
+  updateContact,
+}
